move package computation.dbcleaner to org.sonar.server
there should not be any org.sonar.core package in module sonar-server
